Kernel/Storage: Move Ramdisk code into a separate subdirectory
