+++ Kernel/Storage/Ramdisk/Device.cpp
@@ -8,8 +8,8 @@
 #include <AK/StringView.h>
 #include <Kernel/Devices/DeviceManagement.h>
 #include <Kernel/FileSystem/OpenFileDescription.h>
-#include <Kernel/Storage/RamdiskController.h>
-#include <Kernel/Storage/RamdiskDevice.h>
+#include <Kernel/Storage/Ramdisk/Controller.h>
+#include <Kernel/Storage/Ramdisk/Device.h>
 
 namespace Kernel {
